# server/Dash.Server/Contracts/ElementRequest.cs
using System.ComponentModel.DataAnnotations;
using System;

namespace Dash.Server.Contracts;

public class ElementRequest
{
    [Required]
    public string Config { get; set; } = string.Empty;
    public string? Content { get; set; }
    public DateTime? ExpireDate { get; set; }
    [Required]
    public Guid ComponentId { get; set; }
    [Required]
    public Guid DisplayId { get; set; }
}