# server/Dash.Server/openapi/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Dash",
    "version": "v1.0"
  },
  "paths": {
    "/api/Calendar/{id}": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "operationId": "Calendar_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Calendar"
        ],
        "operationId": "Calendar_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Calendar"
        ],
        "operationId": "Calendar_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Calendar": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "operationId": "Calendar_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Calendar"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Calendar"
        ],
        "operationId": "Calendar_Create",
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Calendar/event/{id}": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "operationId": "Calendar_GetEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEvent"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Calendar"
        ],
        "operationId": "Calendar_UpdateEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarEventRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Calendar"
        ],
        "operationId": "Calendar_DeleteEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Calendar/event": {
      "post": {
        "tags": [
          "Calendar"
        ],
        "operationId": "Calendar_CreateEvent",
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarEventRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Calendar/events": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "operationId": "Calendar_GetEvents",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-position": 1
          },
          {
            "name": "end",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarEvent"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Element": {
      "post": {
        "tags": [
          "Element"
        ],
        "operationId": "Element_Create",
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ElementRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Element/{id}": {
      "post": {
        "tags": [
          "Element"
        ],
        "operationId": "Element_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ElementRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Element/{id}/{displayId}": {
      "delete": {
        "tags": [
          "Element"
        ],
        "operationId": "Element_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "displayId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Element/{displayId}": {
      "get": {
        "tags": [
          "Element"
        ],
        "operationId": "Element_GetByDisplay",
        "parameters": [
          {
            "name": "displayId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Element"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Pinboard": {
      "get": {
        "tags": [
          "Pinboard"
        ],
        "operationId": "Pinboard_List",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Information"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pinboard"
        ],
        "operationId": "Pinboard_Delete",
        "requestBody": {
          "x-name": "ids",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pinboard"
        ],
        "operationId": "Pinboard_Save",
        "requestBody": {
          "x-name": "data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Information"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Pinboard/{id}": {
      "get": {
        "tags": [
          "Pinboard"
        ],
        "operationId": "Pinboard_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Information"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Calendar": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "prodId",
          "name",
          "hexColor",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "prodId": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "hexColor": {
            "type": "string",
            "minLength": 1
          },
          "version": {
            "type": "string",
            "minLength": 1
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalendarEvent"
            }
          }
        }
      },
      "CalendarEvent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "uid",
          "dtStart",
          "dtEnd",
          "dtStamp",
          "calendarId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "uid": {
            "type": "string",
            "minLength": 1
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "class": {
            "type": "string"
          },
          "dtStart": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "dtEnd": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "dtStamp": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "calendarId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "calendar": {
            "$ref": "#/components/schemas/Calendar"
          }
        }
      },
      "CalendarRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "color"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "color": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "CalendarEventRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "dtStart",
          "dtEnd",
          "calendarId"
        ],
        "properties": {
          "location": {
            "type": "string",
            "nullable": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "dtStart": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "dtEnd": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "calendarId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          }
        }
      },
      "Display": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "rows",
          "columns",
          "shareId"
        ],
        "properties": {
          "displayId": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "rows": {
            "type": "integer",
            "format": "int32"
          },
          "columns": {
            "type": "integer",
            "format": "int32"
          },
          "shareId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "elements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Element"
            }
          }
        }
      },
      "Element": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "config",
          "creationDate",
          "componentId",
          "displayId"
        ],
        "properties": {
          "elementId": {
            "type": "string",
            "format": "guid"
          },
          "config": {
            "type": "string",
            "minLength": 1
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "creationDate": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "expireDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "componentId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "component": {
            "$ref": "#/components/schemas/Component"
          },
          "displayId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "display": {
            "$ref": "#/components/schemas/Display"
          }
        }
      },
      "Component": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier",
          "name",
          "config"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "identifier": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "config": {
            "type": "string",
            "minLength": 1
          },
          "elements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Element"
            }
          }
        }
      },
      "ElementRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "config",
          "componentId",
          "displayId"
        ],
        "properties": {
          "config": {
            "type": "string",
            "minLength": 1
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "expireDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "componentId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "displayId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          }
        }
      },
      "Information": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "expiredDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isFocused": {
            "type": "boolean"
          },
          "config": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}